[SPARK-57478][SQL] Extract tar container-writing helpers into a shared trait
writeArchive/archiveExtensions/writeCorruptArchive were duplicated between
TarArchiveReadBase and the standalone TextTarArchiveReadSuite (text can't mix in
TarArchiveReadBase, whose two-column tests don't fit text's single value column).
Move them to a new TarArchiveTestUtils trait that both mix in, so the tar
container logic lives in one place.
